forces install of certifi package on windows
trying to fix issue with Travis CI build failing.  It gets up to the point of running setup.py and then throws the following error: ssl.SSLCertVerificationError: [SSL: CERTIFICATE_VERIFY_FAILED] certificate verify failed: unable to get local issuer certificate (_ssl.c:1108).  Some googling revealed that it is an issue with missing or out of date certs on the computer.  I noticed that the install recommended upgrading pip, so maybe that could have something to do with it.  Once recommendation I saw was to upgrade the certifi package which sounds like python's local store of certificates.