adds pip install on windows builds
Now the tests pass, but deployment fails.  Still the same issue, however.  I noticed that I fix the version of cibuildwheel which I use.  I'll take out the version requirement and just use the most recent one in case it pulls in certs.  I have also added a line to upgrade pip before I do anything.  Not sure if this will fix it, but it can't hurt.